services/processor/requirements.txt
# Web framework
fastapi==0.115.12
uvicorn[standard]==0.34.0

# Database — async PostgreSQL + pgvector
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.40
alembic==1.15.2
pgvector==0.4.1

# Document parsing
pdfplumber==0.11.6
python-docx==1.1.2
pytesseract==0.3.13
pdf2image==1.17.0

# AWS SDK
boto3==1.37.30
botocore==1.37.30

# OpenAI
openai==1.75.0

# Utilities
python-multipart==0.0.20
httpx==0.28.1
pydantic==2.11.3
pydantic-settings==2.9.1
python-dotenv==1.1.0
